view random comic working and can view previous ones by swiping right
