+++ assets/Comic.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 
+
 struct Comic: Codable {
   var month: String
   var num: Int
